[WIP] Some architecture changes

+++ src/zx/machine/specs.rs
@@ -3,29 +3,29 @@ use zx::screen::BORDER_COLS;
 /// Immutable type (Builder is not public in outer module)
 pub struct ZXSpecs {
     // frequencies
-    pub freq_cpu: u64,
+    pub freq_cpu: usize,
     // first_pixel_clocks
-    pub clocks_first_pixel: u64,
+    pub clocks_first_pixel: usize,
     // row clocks
-    pub clocks_left_border: u64,
-    pub clocks_screen_row: u64,
-    pub clocks_right_border: u64,
-    pub clocks_retrace: u64,
-    pub clocks_line: u64,
-    pub clocks_line_base: Vec<u64>,
+    pub clocks_left_border: usize,
+    pub clocks_screen_row: usize,
+    pub clocks_right_border: usize,
+    pub clocks_retrace: usize,
+    pub clocks_line: usize,
+    pub clocks_line_base: Vec<usize>,
     // frame
-    pub clocks_frame: u64,
+    pub clocks_frame: usize,
     // lines metrics
-    pub lines_vsync: u64,
-    pub lines_top_border: u64,
-    pub lines_screen: u64,
-    pub lines_bottom_border: u64,
-    pub lines_all: u64,
+    pub lines_vsync: usize,
+    pub lines_top_border: usize,
+    pub lines_screen: usize,
+    pub lines_bottom_border: usize,
+    pub lines_all: usize,
     // interrupt
-    pub interrupt_length: u64,
+    pub interrupt_length: usize,
     // contention
-    pub contention_offset: u64,
-    pub contention_pattern: [u64; 8],
+    pub contention_offset: usize,
+    pub contention_pattern: [usize; 8],
 }
 
 /// Specs builder, used to make static valiables with machines specs
@@ -72,7 +72,7 @@ impl ZXSpecsBuilder {
         // 4*4 is 4 border columns * 4 clocks per column
         self.specs.clocks_line_base.push(self.specs.clocks_first_pixel -
                                          self.specs.lines_top_border * self.specs.clocks_line -
-                                         BORDER_COLS as u64 * 4);
+                                         BORDER_COLS * 4);
         // + 1 because TStates in calculations may be > frame length (CHECK)
         let lines_count = self.specs.lines_all + 1;
         for _ in 1..lines_count {
@@ -83,12 +83,13 @@ impl ZXSpecsBuilder {
         self.specs
     }
     /// Changes CPU frequency
-    pub fn freq_cpu(mut self, value: u64) -> Self {
+    pub fn freq_cpu(mut self, value: usize) -> Self {
         self.specs.freq_cpu = value;
         self
     }
     /// Changes Clocks per left border, screen render, left border and retrace
-    pub fn clocks_row(mut self, lborder: u64, screen: u64, rborder: u64, retrace: u64) -> Self {
+    pub fn clocks_row(mut self, lborder: usize, screen: usize, rborder: usize, retrace: usize)
+        -> Self {
         self.specs.clocks_left_border = lborder;
         self.specs.clocks_screen_row = screen;
         self.specs.clocks_right_border = rborder;
@@ -97,12 +98,12 @@ impl ZXSpecsBuilder {
         self
     }
     /// Changes first pixel clocks
-    pub fn clocks_first_pixel(mut self, value: u64) -> Self {
+    pub fn clocks_first_pixel(mut self, value: usize) -> Self {
         self.specs.clocks_first_pixel = value;
         self
     }
     /// Changes lines per top border, screen, bottom border and vsync
-    pub fn lines(mut self, tborder: u64, screen: u64, bborder: u64, vsync: u64) -> Self {
+    pub fn lines(mut self, tborder: usize, screen: usize, bborder: usize, vsync: usize) -> Self {
         self.specs.lines_vsync = vsync;
         self.specs.lines_top_border = tborder;
         self.specs.lines_screen = screen;
@@ -111,13 +112,13 @@ impl ZXSpecsBuilder {
         self
     }
     /// Changes contention pattern
-    pub fn contention(mut self, pattern: [u64; 8], offset: u64) -> Self {
+    pub fn contention(mut self, pattern: [usize; 8], offset: usize) -> Self {
         self.specs.contention_pattern = pattern;
         self.specs.contention_offset = offset;
         self
     }
     /// changes interrupt length
-    pub fn interrupt_length(mut self, value: u64) -> Self {
+    pub fn interrupt_length(mut self, value: usize) -> Self {
         self.specs.interrupt_length = value;
         self
     }
